fix(TES): remove from public API internals of PassKey
Equivalent to the change in TokTok/c-toxcore#310 where pass-key
internals are no longer exposed publicly, since there is no need to
expose them.
